Split in packs, provide entry points
