+++ src/packs/fns.js
@@ -1,0 +1,97 @@
+export { default as colorsNamed } from '../colors/named.js';
+export { default as blend } from '../blend.js';
+export { default as random } from '../random.js';
+
+export {
+	fixupHueShorter,
+	fixupHueLonger,
+	fixupHueIncreasing,
+	fixupHueDecreasing
+} from '../fixup/hue.js';
+
+export { fixupAlpha } from '../fixup/alpha.js';
+
+export {
+	mapper,
+	mapAlphaMultiply,
+	mapAlphaDivide,
+	mapTransferLinear,
+	mapTransferGamma
+} from '../map.js';
+
+export { average, averageAngle, averageNumber } from '../average.js';
+
+export { default as round } from '../round.js';
+export {
+	interpolate,
+	interpolateWith,
+	interpolateWithPremultipliedAlpha
+} from '../interpolate/interpolate.js';
+
+export { interpolatorLinear } from '../interpolate/linear.js';
+
+export { interpolatorPiecewise } from '../interpolate/piecewise.js';
+
+export {
+	interpolatorSplineBasis,
+	interpolatorSplineBasisClosed
+} from '../interpolate/splineBasis.js';
+
+export {
+	interpolatorSplineNatural,
+	interpolatorSplineNaturalClosed
+} from '../interpolate/splineNatural.js';
+
+export {
+	interpolatorSplineMonotone,
+	interpolatorSplineMonotone2,
+	interpolatorSplineMonotoneClosed
+} from '../interpolate/splineMonotone.js';
+
+export { default as lerp } from '../interpolate/lerp.js';
+export { default as samples } from '../samples.js';
+export { default as displayable } from '../displayable.js';
+export { clampRgb, clampChroma } from '../clamp.js';
+export { default as nearest } from '../nearest.js';
+export { getModeDefinition } from '../modes.js';
+export { default as parse } from '../parse.js';
+
+export {
+	differenceEuclidean,
+	differenceCie76,
+	differenceCie94,
+	differenceCiede2000,
+	differenceCmc,
+	differenceHyab,
+	differenceHueSaturation,
+	differenceHueChroma,
+	differenceHueNaive
+} from '../difference.js';
+
+export {
+	filterBrightness,
+	filterContrast,
+	filterSepia,
+	filterInvert,
+	filterSaturate,
+	filterGrayscale,
+	filterHueRotate
+} from '../filter.js';
+
+export {
+	filterDeficiencyProt,
+	filterDeficiencyDeuter,
+	filterDeficiencyTrit
+} from '../deficiency.js';
+
+// Easings
+export { default as easingMidpoint } from '../easing/midpoint.js';
+export { default as easingSmoothstep } from '../easing/smoothstep.js';
+export { default as easingSmootherstep } from '../easing/smootherstep.js';
+export { default as easingInOutSine } from '../easing/inOutSine.js';
+export { default as easingGamma } from '../easing/gamma.js';
+
+export {
+	luminance as wcagLuminance,
+	contrast as wcagContrast
+} from '../wcag.js';
